# wafir_project_export/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Wafir API
  description: API for Wafir - Sharia-Compliant Personal Finance & Investment App
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server

paths:
  /auth/login:
    post:
      summary: User Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Successful Login
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  /portfolio/summary:
    get:
      summary: Get Portfolio Summary
      tags: [Portfolio]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Portfolio summary data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortfolioSummary'

  /transactions:
    get:
      summary: Get Transactions
      tags: [Transactions]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
    post:
      summary: Create Transaction
      tags: [Transactions]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionInput'
      responses:
        '201':
          description: Transaction created

  /zakat/calculate:
    get:
      summary: Calculate Estimated Zakat
      tags: [Zakat]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Zakat calculation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalZakatDue: { type: number }
                  currency: { type: string }
                  breakdown:
                    type: array
                    items:
                      type: object
                      properties:
                        assetId: { type: string }
                        amount: { type: number }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    AuthResponse:
      type: object
      properties:
        accessToken: { type: string }
        user:
          $ref: '#/components/schemas/User'

    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        name: { type: string }

    PortfolioSummary:
      type: object
      properties:
        totalValue: { type: number }
        currency: { type: string }
        allocation:
          type: object
          properties:
            cash: { type: number }
            equity: { type: number }
            realEstate: { type: number }
            sukuk: { type: number }

    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        amount: { type: number }
        description: { type: string }
        date: { type: string, format: date-time }
        category: { type: string }
        type: { type: string, enum: [INCOME, EXPENSE] }

    TransactionInput:
      type: object
      properties:
        amount: { type: number }
        description: { type: string }
        date: { type: string, format: date-time }
        categoryId: { type: string, format: uuid }